Used recyclerView for display and also put entries in database
This mostly works. When app initially runs it looks good. But
when I scroll it's like there are 2 sets of characters initially
right on top of each other. When I scroll, one follows the
scrolling and the other stays so every line is doubled as one moves
and the other stays.