Move the tracking to the ViewModel.
I was using the chronometer and button values as state info. When
the UI had to restart like when canging device orientation that
state was lost. So, I moved most of the state tracking to the
ViweModel. But I still need to save the reports and difference and
gain. I have to make sure I have enough info to put the state back.
Then when View is created I need to initialize the buttons and
chronometers and reports and difference and gain to the values from
the Fragment with values stored in ViewModel.